refactor: viewmodel 분리
